+++ swan/dataset/__init__.py
@@ -1,5 +1,5 @@
 from .fingerprints_data import FingerprintsData
-from .graph_data import GraphData
+from .torch_geometric_graph_data import TorchGeometricGraphData
+from .dgl_graph_data import DGLGraphData
 
-
-__all__ = ["FingerprintsData", "GraphData"]
+__all__ = ["DGLGraphData", "FingerprintsData", "TorchGeometricGraphData"]