+++ swan/modeller/models/__init__.py
@@ -3,5 +3,6 @@
 from .fingerprint_models import FingerprintFullyConnected
 from .graph_models import MPNN
 from .equivariant_models import InvariantPolynomial
+from .se3_transformer import TFN, SE3Transformer
 
-__all__ = ["FingerprintFullyConnected", "InvariantPolynomial", "MPNN"]
+__all__ = ["FingerprintFullyConnected", "InvariantPolynomial", "MPNN", "SE3Transformer", "TFN"]